implement Submit Dataset Proposal workflow

+++ api/models/repositories.go
@@ -9,7 +9,7 @@ type Repository struct {
 	OrganizationNodeId string     `dynamodbav:"OrganizationNodeId"`
 	Name               string     `dynamodbav:"Name"`
 	DisplayName        string     `dynamodbav:"DisplayName"`
-	RepositoryId       int64      `dynamodbav:"RepositoryId"`
+	RepositoryId       int        `dynamodbav:"RepositoryId"`
 	Type               string     `dynamodbav:"Type"`
 	Description        string     `dynamodbav:"Description"`
 	URL                string     `dynamodbav:"URL"`
